feat: deploy and verify

+++ contracts/ClassicalNFT.sol
@@ -89,6 +89,7 @@ contract ClassicalNFT is
     event SetMintPrice(uint256 indexed price);
 
     // name and symbol
+
     constructor(
         address _treasuryAddress,
         address _publicGoodAddress,
@@ -118,7 +119,7 @@ contract ClassicalNFT is
         bytes memory signature,
         string memory _bookId,
         address classicalNFTAddr,
-        uint chainId
+        uint256 chainId
     ) public view returns (bool) {
         bytes32 data = keccak256(
             abi.encodePacked(msg.sender, _bookId, classicalNFTAddr, chainId)
@@ -129,9 +130,10 @@ contract ClassicalNFT is
     }
 
     //remove addresses to whiteList
-    function removeFromWhitelist(
-        address[] calldata _addrArr
-    ) external onlyOwner {
+    function removeFromWhitelist(address[] calldata _addrArr)
+        external
+        onlyOwner
+    {
         for (uint256 i = 0; i < _addrArr.length; i++) {
             _revokeRole(FREE_MINT_ROLE, _addrArr[i]);
         }
@@ -187,7 +189,7 @@ contract ClassicalNFT is
         string memory _bookId,
         bytes memory signature,
         address classicalNFTAddr,
-        uint chainId
+        uint256 chainId
     ) public payable nonReentrant returns (uint256) {
         require(isMintEnabled, "Minting not enabled");
         require(currentSupply <= maxSupply, "Max supply reached");
@@ -222,7 +224,7 @@ contract ClassicalNFT is
         string memory _bookId,
         bytes memory signature,
         address classicalNFTAddr,
-        uint chainId
+        uint256 chainId
     ) public nonReentrant onlyWhiteList returns (uint256) {
         require(isMintEnabled, "Minting not enabled");
         require(currentSupply <= maxSupply, "Max supply reached");
@@ -251,10 +253,12 @@ contract ClassicalNFT is
         return bookIds;
     }
 
-    function mintReserve(
-        address _recipient,
-        string memory _bookId
-    ) external nonReentrant onlyOwner returns (uint256) {
+    function mintReserve(address _recipient, string memory _bookId)
+        external
+        nonReentrant
+        onlyOwner
+        returns (uint256)
+    {
         require(isMintEnabled, "Minting not enabled");
         require(currentSupply <= maxSupply, "Max supply reached");
         require(!bookList[_bookId], "Book id exist");
@@ -295,9 +299,12 @@ contract ClassicalNFT is
 
     /// @dev Overridden in order to make it an onlyOwner function
     // TODO: ERC20 的 transfer
-    function withdrawPayments(
-        address payable
-    ) public virtual override onlyOwner {
+    function withdrawPayments(address payable)
+        public
+        virtual
+        override
+        onlyOwner
+    {
         // 先取
         super.withdrawPayments(payable(address(this)));
 
@@ -312,16 +319,19 @@ contract ClassicalNFT is
         payable(treasuryAddress).transfer(address(this).balance);
     }
 
-    function supportsInterface(
-        bytes4 interfaceId
-    ) public view override(ERC721, ERC2981, AccessControl) returns (bool) {
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        override(ERC721, ERC2981, AccessControl)
+        returns (bool)
+    {
         return super.supportsInterface(interfaceId);
     }
 
-    function setTokenRoyalty(
-        uint256 tokenId,
-        uint96 feeNumerator
-    ) public onlyOwner {
+    function setTokenRoyalty(uint256 tokenId, uint96 feeNumerator)
+        public
+        onlyOwner
+    {
         // TODO: 直接到国库，会导致 donate 无法计算，_asynctransfer 的地址和 receiver 不一样？
         _setTokenRoyalty(tokenId, address(this), feeNumerator);
     }
@@ -335,9 +345,10 @@ contract ClassicalNFT is
         treasuryAddress = _treasuryAddress;
     }
 
-    function setPublicGoodAddress(
-        address _publicGoodAddress
-    ) external onlyOwner {
+    function setPublicGoodAddress(address _publicGoodAddress)
+        external
+        onlyOwner
+    {
         publicGoodAddress = _publicGoodAddress;
     }
